refactor: formatStat を lib/format.ts に移動し worktree を test exclude に追加
- formatStat の定義元を NaGrandTotalTable.tsx → lib/format.ts に変更
- NaCrossTable.tsx の import を lib/format からの統合 import に修正
- .claude/** を vitest exclude に追加（worktree 内テストの誤検出防止）
