+++ src/components/aggregation/NaGrandTotalTable.tsx
@@ -1,4 +1,5 @@
 import type { Tally } from "../../lib/agg/types";
+import { formatStat } from "../../lib/format";
 import { t } from "../../lib/i18n";
 import { Th, Td } from "./TableCells";
 
@@ -32,19 +33,4 @@ export function NaGrandTotalTable({ tally }: NaGrandTotalTableProps) {
       </tbody>
     </table>
   );
-}
-
-export function formatStat(key: string, value: number): string {
-  if (key === "n") return value.toLocaleString();
-  return value.toFixed(2);
-}
-
-if (import.meta.vitest) {
-  const { test, expect } = import.meta.vitest;
-  test("formatStat: n はロケール整数表示", () => {
-    expect(formatStat("n", 1234)).toBe("1,234");
-  });
-  test("formatStat: mean は小数2桁", () => {
-    expect(formatStat("mean", 3.5)).toBe("3.50");
-  });
 }
